Refactor Multi/Single Service Tests into BuildRecalculations and GetAvailableSlots tests.
(cherry picked from commit de744612aa9a8d8b283db55b2acb27b19dad22dd)
